Users delete logic added

+++ users-service/src/main/java/co/com/poli/usersservice/service/UserServiceImpl.java
@@ -1,5 +1,7 @@
 package co.com.poli.usersservice.service;
 
+import co.com.poli.usersservice.clientFeign.BookingClient;
+import co.com.poli.usersservice.commons.Functions;
 import co.com.poli.usersservice.persistance.entity.User;
 import co.com.poli.usersservice.persistance.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -13,6 +15,8 @@ import java.util.List;
 public class UserServiceImpl implements UserService{
 
     private final UserRepository userRepository;
+
+    private final BookingClient bookingClient;
 
     @Override
     @Transactional(readOnly = true)
@@ -28,8 +32,14 @@ public class UserServiceImpl implements UserService{
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public void delete(String id) {
-        userRepository.deleteById(Long.valueOf(id));
+    public Boolean delete(String id) {
+        boolean wasDeleted = false;
+        List<?> bookingList = Functions.convertObjectToList(bookingClient.findByUserID(id).getData());
+        if(bookingList == null || bookingList.isEmpty()){
+            userRepository.deleteById(Long.valueOf(id));
+            wasDeleted = true;
+        }
+        return wasDeleted;
     }
 
     @Override
@@ -37,4 +47,5 @@ public class UserServiceImpl implements UserService{
     public User findById(String id) {
         return userRepository.findById(Long.valueOf(id)).orElse(null);
     }
+
 }
